+++ compiler/src/iree/compiler/Codegen/Common/GPU/GPUPatterns.h
@@ -29,6 +29,10 @@ void populateContractPromotionPatterns(RewritePatternSet &patterns,
 
 void populateDropSharedMemoryDeallocOpPatterns(RewritePatternSet &patterns);
 
+/// Adds patterns to convert stretching broadcasts (broadcasting a non-unit dim
+/// from 1) into broadcast + transpose so that layout analysis can handle them.
+void populateVectorLayoutCanonicalizations(RewritePatternSet &patterns);
+
 void populateGPUDistributionPatterns(RewritePatternSet &patterns);
 
 void populateGPUDistributeNestedLayoutAttrPatterns(
